[Desktop] rename api service.js to httpService.js update dependencies add missing imports
