+++ DesktopReact/frontend/src/api/httpService.js
@@ -1,14 +1,20 @@
 const baseURL = 'http://localhost:33248';
-const GET = 'GET';
-const POST = 'POST';
-const DELETE = 'DELETE';
-const UPDATE = 'UPDATE';
+const GET = 'GET', POST = 'POST', DELETE = 'DELETE', UPDATE = 'UPDATE', PUT = 'PUT';
 const STATUS_OK = 'Ok';
+const headers = new Headers({'Accept': 'application/json', 'Content-Type': 'application/json'});
 
 function followUser() {
 
 }
+
 function unfollowUser() {
+
+}
+
+function reactToPost(body) {
+    return {
+        status: 'ok', "reactions": [[0, 0], [1, 0], [2, 0], [3, 0], [4, 0], [5, 0], [6, 0]]
+    }
 
 }
 
@@ -25,7 +31,8 @@ function getPosts(postBody) {
 }
 
 function call(endpoint, method, body) {
-    return fetch(baseURL + endpoint, {method: method, body: body})
+    console.log(body)
+    return fetch(baseURL + endpoint, {method: method, body: body, headers: headers, mode: 'cors'})
         .then(response => response)
         .catch(error => error)
 }
