+++ DesktopReact/frontend/src/components/postPlaceholder/PostPlaceholder.js
@@ -2,14 +2,14 @@ import "./PostPlaceholder.css"
 import React, {useEffect, useState} from 'react';
 import {Alert, Avatar, Button, Card, Snackbar, Stack, TextField, Typography} from "@mui/material";
 import pic from "../../images/superalgos.png"
-import {createPost, STATUS_OK} from "../../api/service";
+import {createPost, STATUS_OK} from "../../api/httpService";
 
 const PostPlaceholder = () => {
     const [postText, setPostText] = useState(undefined);
     const [open, setOpen] = useState(false);
 
     const onButtonClick = async () => {
-        let {status} = await createPost({body: postText}).then(response => response.json());
+        let {status} = await createPost({body: postText}).then(response => console.log(response));
         if (status === STATUS_OK) return;
         setOpen(true);
