+++ DesktopReact/frontend/src/components/suggestedUsers/SuggestedUsers.js
@@ -2,7 +2,7 @@ import "./SuggestedUsers.css"
 import React, {useEffect, useState} from 'react';
 import ShowMoreUsers from "../showMoreUsers/ShowMoreUsers";
 import {Stack} from "@mui/material";
-import {getProfiles, STATUS_OK} from "../../api/service";
+import {getProfiles, STATUS_OK} from "../../api/httpService";
 import UserCard from "../User/UserCard";
