+++ package.json
@@ -43,6 +43,8 @@
     "case-sensitive-paths-webpack-plugin": "2.3.0",
     "ccxt": "^1.61.93",
     "child_process": "^1.0.2",
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
     "css-loader": "4.3.0",
     "discord-api-types": "^0.25.2",
     "discord.js": "^13.3.1",
